# Optimize translation for ListWorkflowExecutions and GetWorkflowExecutionRawHistoryV2