--- a/interceptor/reflection.go
+++ b/interceptor/reflection.go
@@ -9,8 +9,8 @@ import (
 	"go.temporal.io/api/common/v1"
 	"go.temporal.io/api/enums/v1"
 	"go.temporal.io/api/history/v1"
-	"go.temporal.io/api/workflowservice/v1"
 	"go.temporal.io/api/namespace/v1"
+	"go.temporal.io/api/workflowservice/v1"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
 	"go.temporal.io/server/common/persistence/serialization"
@@ -136,7 +136,7 @@ type visitor func(logger log.Logger, obj any, match stringMatcher) (bool, error)
 // in the given object. When it finds namespace string fields, it invokes
 // the provided match function.
 func visitNamespace(logger log.Logger, obj any, match stringMatcher) (bool, error) {
-	if isSkippableForNamespaceTranslation(logger, obj) {
+	if isSkippableForNamespaceTranslation(obj) {
 		return false, nil
 	}
 
@@ -164,6 +164,16 @@ func visitNamespace(logger log.Logger, obj any, match stringMatcher) (bool, erro
 				info.Name = newName
 			}
 			matched = matched || ok
+		} else if hist, ok := vwp.Interface().(*history.History); ok && hist != nil {
+			for _, evt := range hist.GetEvents() {
+				// Do the recursive call here so that we check `isSkippableForNamespaceTranslation`.
+				m, err := visitNamespace(logger, evt, match)
+				matched = matched || m
+				if err != nil {
+					return visit.Stop, err
+				}
+			}
+			return visit.Continue, nil
 		} else if dataBlobFieldNames[fieldType.Name] {
 			changed, err := visitDataBlobs(logger, vwp, match, visitNamespace)
 			matched = matched || changed
@@ -409,7 +419,7 @@ func validateAndRepairHistoryEvents(events []*history122.HistoryEvent) (bool, er
 	return changed, nil
 }
 
-func isSkippableForNamespaceTranslation(logger log.Logger, vAny any) bool {
+func isSkippableForNamespaceTranslation(vAny any) (result bool) {
 	switch v := vAny.(type) {
 	case *workflowservice.ListWorkflowExecutionsResponse:
 		return true
@@ -424,13 +434,21 @@ func isSkippableForNamespaceTranslation(logger log.Logger, vAny any) bool {
 			// Any events with a namespace fields should not skip translation.
 			_, skippable := namespaceTranslationSkippableHistoryEvents[evt.GetEventType()]
 			if !skippable {
-				logger.Info("unskippable history event type", tag.NewAnyTag("type", evt.GetEventType()))
-
 				return false
 			}
 		}
 		// Only skippable if all events in the list are skippable
 		return true
+	case *history.HistoryEvent:
+		// If this namespace field is set, do not skip translation.
+		for _, l := range v.Links {
+			if len(l.GetWorkflowEvent().GetNamespace()) > 0 {
+				return false
+			}
+		}
+		// Any events with a namespace fields should not skip translation.
+		_, skippable := namespaceTranslationSkippableHistoryEvents[v.GetEventType()]
+		return skippable
 	}
 
 	return false