## What was changed

Optimize namespace translation in a couple of cases where we've seen high latency.

1. workflowservice.ListWorkflowExecutionsResponse - we can skip namespace translation for this type
2. adminservice.GetWorkflowExecutionRawHistoryV2Response - we can skip translation for most history event types. With a 20MB test workflow, this improves latency through the proxy from about 3s to about 1.6s.
3. workflowservice.GetWorkflowExecutionHistoryResponse - similar to (2). I see an improvement of about 2.5s to 1s on a 20MB test workflow.

## Why?
<!-- Tell your future self why have you made these changes -->

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

4. How was this tested:
<!--- Please describe how you tested your changes/how we can test them -->

5. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->